requirements.txt
faiss-cpu
python-docx
streamlit
python-dotenv
openai>=1.0.0
bcrypt
pandas
numpy

